Move backend into a pip package
