+++ back/setup.py
@@ -1,0 +1,3 @@
+from setuptools import setup, find_packages
+
+setup(name="tmeit-backend", packages=find_packages())
